Add meta-freescale
This layer gives us support for a selection of boards based on NXP
(Freescale) processors.  This is normally a good starting point for
bringing up a board as our board will in all likelihood be based on one
of these reference designs.

Later, when our project matures, we can just get our own kernel and
bootloader into place and then this layer might no longer be necessary.