Add firmware for wifi module
Normally, this would go into a "machine" file; currently, however, we
are piggy-backing on NXP EVK's so we'll just add this to the
packagegroup for now.

In the future, this should be:

MACHINE_EXTRA_RDEPENDS = "linux-firmware-sd8887"

in some solarmachine.conf file.  That's for the future, though, when SA
has its own hardware ready.