Add machine solarevk
The primary purpose of this is to switch to the mainline kernel.
Longterm, solar will have it's own platform that's not just a modified
EVK and this machine configuration can be evolved at that time to
support that platform.  Initially, though, this is just NXP's platform
with a custom kernel.